pyproject.toml
[tool.poetry]
name = "bang"
version = "0.1.0"
description = ""
authors = ["goeasyonng <meoa1026@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
drf-yasg = "^1.21.5"
daphne = "^4.0.0"
channels = "^4.0.0"
django-seed = "^0.3.1"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
